+++ src/code.js/main.js
@@ -1,7 +1,7 @@
 import SearchApiTrend from "./apiTrendService.js";
 
 import trendMovieTpl from '../templates/trendfilm-cards.hbs';
-import articleTpl from '../templates/modal-card.hbs'
+import articleTpl from '../templates/modal-card.hbs';
 
 const refs = {
     trendContainer: document.querySelector('.js-trend-list'),
@@ -19,7 +19,7 @@ const refs = {
     btnListPage: document.querySelector('.button-list__page'),
     wrapperEl: document.querySelector('.modal-movie-wrapper'),
     poster: document.querySelector('.card-poster-wrapper'),
-    information: document.querySelector('.modal-movie-information')
+    information: document.querySelector('.modal-movie-information'),
 
     btnAddWatched: document.querySelector('[data-name="watched"]'),
     btnAddQueue: document.querySelector('[data-name="queue"]'),
@@ -30,6 +30,7 @@ export default refs;
 SearchApiTrend.fetchtrend().then(results => {
     renderMovies(results)
 });
+
 
 function renderMovies(results) {
     // console.log(results);
@@ -122,9 +123,6 @@ refs.closeModalBtn.addEventListener('click',onBtnClose)
 function onBtnClose(){
 
 
-
-
-
 refs.trendContainer.addEventListener('click',e => {
     if(e.target.nodeName !=='IMG'){
         return
@@ -197,83 +195,84 @@ refs.overlay.addEventListener('click', evt => {
     openCloseModal();
    }
 })
+}
 
 
 
 ///////////////////////начало пагинации 
-const API__KEY = '44d74a10460e9a32f8546bed31d47780';
-const BASE__URL = 'https://api.themoviedb.org/3/discover/';
-let n = 1;
-refs.next.addEventListener('click',e => {
-    console.log(e.target)
-    n += 1;
-     if(n > 500){
-         n--
-         return
-     }
-     btnCreate()
-    return feachMuvie()
+// const API__KEY = '44d74a10460e9a32f8546bed31d47780';
+// const BASE__URL = 'https://api.themoviedb.org/3/discover/';
+// let n = 1;
+// refs.next.addEventListener('click',e => {
+//     console.log(e.target)
+//     n += 1;
+//      if(n > 500){
+//          n--
+//          return
+//      }
+//      btnCreate()
+//     return feachMuvie()
 
-})
+// })
 
-if(n === 1){
- refs.btnListPage.classList.add('button-list__page--current');
-}
+// if(n === 1){
+//  refs.btnListPage.classList.add('button-list__page--current');
+// }
 
-refs.pr.addEventListener('click', e =>{
-    n-=1;
-    if(n === 0){
-        n++
-        return
-    }
-    btnCreate()
-    return feachMuvie()
+// refs.pr.addEventListener('click', e =>{
+//     n-=1;
+//     if(n === 0){
+//         n++
+//         return
+//     }
+//     btnCreate()
+//     return feachMuvie()
 
-})
-refs.btnList.addEventListener('click', e => {
+// })
+// refs.btnList.addEventListener('click', e => {
 
     
-    if(e.target.nodeName!== 'BUTTON'){
-        return
-    }
+//     if(e.target.nodeName!== 'BUTTON'){
+//         return
+//     }
     
-    console.log(e.target.nodeName,e.currentTarget)
+//     console.log(e.target.nodeName,e.currentTarget)
     
-    n = e.target.textContent -0;
-    btnCreate()
-    feachMuvie()
+//     n = e.target.textContent -0;
+//     btnCreate()
+//     feachMuvie()
 
-})
+// })
 
-function kekw(kuda, n){
-	refs.btnList.insertAdjacentHTML(kuda,`<li class="button-list__item"><button class="button-list__page">${n}</button></li>`)
-}
-function btnCreate(){
-    refs.btnList.innerHTML = '';
-    refs.btnList.insertAdjacentHTML('afterbegin',`<li class="button-list__item button-list__item--curretn"><button class="button-list__page button-list__page--current">${n}</button></li>`)
-    for(let i = 1; i < 3; i++){
-    if(n+i < 501)
-      kekw('beforeend', n+i);
-     if(n-i > 0)
-            kekw('afterbegin', n-i);
-    }
+// function kekw(kuda, n){
+// 	refs.btnList.insertAdjacentHTML(kuda,`<li class="button-list__item"><button class="button-list__page">${n}</button></li>`)
+// }
+// function btnCreate(){
+//     refs.btnList.innerHTML = '';
+//     refs.btnList.insertAdjacentHTML('afterbegin',`<li class="button-list__item button-list__item--curretn"><button class="button-list__page button-list__page--current">${n}</button></li>`)
+//     for(let i = 1; i < 3; i++){
+//     if(n+i < 501)
+//       kekw('beforeend', n+i);
+//      if(n-i > 0)
+//             kekw('afterbegin', n-i);
+//     }
 
-}
+// }
 
-function feachMuvie(){
-    fetch(`${BASE__URL}movie?api_key=${API__KEY}&language=en-US&sort_by=popularity.desc&include_adult=false&include_video=false&page=${n}&with_watch_monetization_types=flatrate`)
-    .then(r => r.json())
-    .then( films => {
-        console.log(films)
-        return films
-    })
-    .then( ({results}) =>{
+// function feachMuvie(){
+//     fetch(`${BASE__URL}movie?api_key=${API__KEY}&language=en-US&sort_by=popularity.desc&include_adult=false&include_video=false&page=${n}&with_watch_monetization_types=flatrate`)
+//     .then(r => r.json())
+//     .then( films => {
+//         console.log(films)
+//         return films
+//     })
+//     .then( ({results}) =>{
 
-        const markUp = trendMovieTpl(results);
-        refs.trendContainer.innerHTML = ''
-        refs.trendContainer.insertAdjacentHTML('beforeend',markUp);
-    })
-} 
+//         const markUp = trendMovieTpl(results);
+//         refs.trendContainer.innerHTML = ''
+//         refs.trendContainer.insertAdjacentHTML('beforeend',markUp);
+//     })
+// } 
 
 //конец пагинации
