+++ src/index.js
@@ -1,4 +1,5 @@
 import './sass/main.scss';
+import './code.js/pagination';
 import * as basicLightbox from 'basiclightbox';
 import 'basiclightbox/dist/basicLightbox.min.css';
 import './code.js/main';
