+++ src/sass/components/_modal.scss
@@ -183,7 +183,7 @@ img {
   font-size: 20px;
   line-height: 1.15;
   text-transform: uppercase;
-}
+
 
   width: 240px;
   height: auto;
@@ -451,6 +451,7 @@ img {
     text-transform: uppercase;
     cursor: pointer;
   }
+}
 }
 
 .modal-left-button {
